ec2-stack/services/worker/requirements.txt
paho-mqtt==2.1.0
influxdb-client==1.46.0
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
pydantic-settings==2.5.2
